Update infrastructure from other Fatiando projects
Add README and other .md files in a standard format, move to a src
layout, add our Makefile, add the doc template, add default CI, update
pyproject.toml to the default.
