+++ .gitignore
@@ -1,25 +1,19 @@
 *.pyc
-*.h5
-*~
-*.so
 *.pyd
+*.so
+.*.swp
+*~
+*.h5
+.cache
 .pytest_cache/
+.eggs
+*.egg-info
+.coverage
+.coverage.*
+.ipynb_checkpoints
+MANIFEST
 build/
 dist/
-.*.swp
-.coverage
-.cache
+**/_version.py
 doc/_build
 doc/api/generated
-doc/gallery
-doc/tutorials
-doc/sample_data
-.ipynb_checkpoints
-*.eggs
-*.egg-info
-MANIFEST
-dask-worker-space
-.coverage.*
-*.ipynb
-**/_version_generated.py
-sg_execution_times.rst
