+++ CITATION.rst
@@ -1,0 +1,8 @@
+Citing Tremelique
+=================
+
+This is research software **made by scientists**. Citations help us justify the
+effort that goes into building and maintaining this project.
+
+If you used it in your research, please consider citing the latest Zenodo
+archive: https://doi.org/ADD_ZENODO_DOI_HERE
