+++ LICENSE.txt
@@ -1,4 +1,4 @@
-Copyright (c) 2024 The Tremelique Developers
+Copyright (c) 2025 The Bordado Developers
 All rights reserved.
 
 Redistribution and use in source and binary forms, with or without modification,
