+++ README.md
@@ -1,0 +1,72 @@
+<h1 align="center">Tremelique</h1>
+
+<p align="center"><strong>Educational 2D elastic wave simulator in Python</strong></p>
+
+<p align="center">
+<a href="https://www.fatiando.org/tremelique"><strong>Documentation</strong> (latest)</a> •
+<a href="https://www.fatiando.org/tremelique/dev"><strong>Documentation</strong> (main branch)</a> •
+<a href="https://github.com/fatiando/tremelique/blob/main/CONTRIBUTING.md"><strong>Contributing</strong></a> •
+<a href="https://www.fatiando.org/contact/"><strong>Contact</strong></a> •
+<a href="https://github.com/orgs/fatiando/discussions"><strong>Ask a question</strong></a>
+</p>
+
+<p align="center">
+Part of the
+<a href="https://www.fatiando.org"><strong>Fatiando a Terra</strong></a>
+project.
+</p>
+
+<p align="center">
+<a href="https://pypi.python.org/pypi/tremelique"><img src="http://img.shields.io/pypi/v/tremelique.svg?style=flat-square" alt="Latest version on PyPI"></a>
+<a href="https://github.com/conda-forge/tremelique-feedstock"><img src="https://img.shields.io/conda/vn/conda-forge/tremelique.svg?style=flat-square" alt="Latest version on conda-forge"></a>
+<a href="https://pypi.python.org/pypi/tremelique"><img src="https://img.shields.io/pypi/pyversions/tremelique.svg?style=flat-square" alt="Compatible Python versions."></a>
+</p>
+
+## About
+
+TBD.
+
+## Project goals
+
+TBD.
+
+## Project status
+
+**Tremelique is in very early stages of development.**
+
+**We welcome feedback and ideas!** This is a great time to bring new ideas on
+how we can improve the project.
+[Join the conversation](https://www.fatiando.org/contact) or submit
+[issues on GitHub](https://github.com/fatiando/tremelique/issues).
+
+## Getting involved
+
+🗨️ **Contact us:**
+Find out more about how to reach us at
+[fatiando.org/contact](https://www.fatiando.org/contact/).
+
+👩🏾‍💻 **Contributing to project development:**
+Please read our
+[Contributing Guide](https://github.com/fatiando/tremelique/blob/main/CONTRIBUTING.md)
+to see how you can help and give feedback.
+
+🧑🏾‍🤝‍🧑🏼 **Code of conduct:**
+This project is released with a
+[Code of Conduct](https://github.com/fatiando/community/blob/main/CODE_OF_CONDUCT.md).
+By participating in this project you agree to abide by its terms.
+
+> **Imposter syndrome disclaimer:**
+> We want your help. **No, really.** There may be a little voice inside your
+> head that is telling you that you're not ready, that you aren't skilled
+> enough to contribute. We assure you that the little voice in your head is
+> wrong. Most importantly, **there are many valuable ways to contribute besides
+> writing code**.
+>
+> *This disclaimer was adapted from the*
+> [MetPy project](https://github.com/Unidata/MetPy).
+
+## License
+
+This is free software: you can redistribute it and/or modify it under the terms
+of the **BSD 3-clause License**. A copy of this license is provided in
+[`LICENSE.txt`](https://github.com/fatiando/tremelique/blob/main/LICENSE.txt).
