+++ doc/Makefile
@@ -25,9 +25,6 @@ clean:
 	rm -rf $(BUILDDIR)/doctrees
 	rm -rf $(BUILDDIR)/linkcheck
 	rm -rf api/generated
-	rm -rf gallery
-	rm -rf tutorials
-	rm -rf sample_data
 	rm -rf .ipynb_checkpoints
 
 api:
