+++ doc/_templates/autosummary/class.rst
@@ -4,29 +4,41 @@
 
 .. autoclass:: {{ objname }}
 
-.. rubric:: Methods Summary
+{% if attributes %}
+
+Attributes
+----------
+
+{% for item in attributes %}
+
+.. autoattribute:: {{ objname }}.{{ item }}
+
+{% endfor %}
+
+{% endif %}
+
+Methods
+-------
+
+.. rubric:: List of methods
 
 .. autosummary::
-    {% for item in members %}
-    {% if item in ['__call__'] %}
-        {{ objname }}.{{ item }}
-    {% endif %}
-    {% endfor %}
     {% for item in methods %}
-    {% if item != '__init__' %}
+    {% if item != "__init__" %}
         {{ objname }}.{{ item }}
     {% endif %}
     {% endfor %}
 
-----
+.. rubric:: Methods documentation
 
 {% for item in methods %}
 {% if item != '__init__' %}
 .. automethod:: {{ objname }}.{{ item }}
+
+----
+
 {% endif %}
 {% endfor %}
-
-.. include:: backreferences/{{ fullname }}.examples
 
 .. raw:: html
