+++ doc/_templates/autosummary/exception.rst
@@ -4,7 +4,6 @@
 
 .. autoexception:: {{ objname }}
 
-
 .. raw:: html
 
      <div style='clear:both'></div>
