+++ doc/_templates/autosummary/function.rst
@@ -4,8 +4,6 @@
 
 .. autofunction:: {{ objname }}
 
-.. include:: backreferences/{{ fullname }}.examples
-
 .. raw:: html
 
      <div style='clear:both'></div>
