+++ doc/changes.rst
@@ -3,10 +3,3 @@
 Changelog
 =========
 
-Version x.y.z
--------------
-
-Released on: YYYY/MM/DD
-
-DOI: https://doi.org/
-
