+++ doc/citing.rst
@@ -1,0 +1,3 @@
+.. _citing:
+
+.. include:: ../CITATION.rst
