+++ doc/conf.py
@@ -1,28 +1,42 @@
+# Copyright (c) 2025 The Tremelique Developers.
+# Distributed under the terms of the BSD 3-Clause License.
+# SPDX-License-Identifier: BSD-3-Clause
+#
+# This code is part of the Fatiando a Terra project (https://www.fatiando.org)
+#
+# Copyright (c) 2025 The tremelique Developers.
+# Distributed under the terms of the BSD 3-Clause License.
+# SPDX-License-Identifier: BSD-3-Clause
+#
+# This code is part of the Fatiando a Terra project (https://www.fatiando.org)
+#
+"""
+Configuration for sphinx https://www.sphinx-doc.org
+"""
+
 import datetime
 
 import tremelique
 
 # Project information
 # -----------------------------------------------------------------------------
-project = "Tremelique"
-copyright = f"{datetime.date.today().year}, The {project} Developers"  # noqa: A001
-if len(tremelique.__version__.split("+")) > 1 or tremelique.__version__ == "unknown":
-    version = "dev"
-else:
-    version = tremelique.__version__
+project = "tremelique"
+copyright = f"{datetime.date.today().year}, The {project} Developers"
+is_dev_version = len(tremelique.__version__.split(".")) > 3
+version = "dev" if is_dev_version else tremelique.__version__
 
 # General configuration
 # -----------------------------------------------------------------------------
 extensions = [
     "sphinx.ext.autodoc",
     "sphinx.ext.autosummary",
+    "numpydoc",
     "sphinx.ext.coverage",
     "sphinx.ext.mathjax",
     "sphinx.ext.doctest",
     "sphinx.ext.viewcode",
     "sphinx.ext.extlinks",
     "sphinx.ext.intersphinx",
-    "sphinx.ext.napoleon",
     "sphinx_design",
     "sphinx_copybutton",
     "jupyter_sphinx",
@@ -33,27 +47,23 @@ extensions = [
 intersphinx_mapping = {
     "python": ("https://docs.python.org/3/", None),
     "numpy": ("https://numpy.org/doc/stable/", None),
-    "xarray": ("https://docs.xarray.dev/en/stable/", None),
-    "matplotlib": ("https://matplotlib.org/stable/", None),
 }
 
 # Autosummary pages will be generated by sphinx-autogen instead of sphinx-build
 autosummary_generate = []
 
-# Otherwise, the Return parameter list looks different from the Parameters list
-napoleon_use_rtype = False
-# Otherwise, the Attributes parameter list looks different from the Parameters
-# list
-napoleon_use_ivar = True
+# Create cross-references for the parameter types in the Parameters, Other
+# Returns and Yields sections of the docstring
+numpydoc_xref_param_type = True
 
-# Always show the source code that generates a plot
-plot_include_source = True
-plot_formats = ["png"]
+# Format the Attributes section like the Parameters section.
+numpydoc_attributes_as_param_list = True
 
 # Sphinx project configuration
 templates_path = ["_templates"]
 exclude_patterns = ["_build", "**.ipynb_checkpoints"]
 source_suffix = ".rst"
+# The encoding of source files
 source_encoding = "utf-8"
 master_doc = "index"
 pygments_style = "default"
@@ -64,17 +74,17 @@ add_function_parentheses = False
 html_title = f'{project} <span class="project-version">{version}</span>'
 # Don't use the logo since it gets in the way of the project name and is
 # repeated in the front page.
-# html_logo = ""
+# html_logo = "_static/tremelique-logo.png"
 html_favicon = "_static/favicon.png"
 html_last_updated_fmt = "%b %d, %Y"
 html_copy_source = True
 html_static_path = ["_static"]
+# CSS files are relative to the static path
+html_css_files = ["custom.css"]
 html_extra_path = []
 html_show_sourcelink = False
 html_show_sphinx = True
 html_show_copyright = True
-# CSS files are relative to the static path
-html_css_files = ["custom.css"]
 
 html_theme = "sphinx_book_theme"
 html_theme_options = {
