+++ doc/versions.rst
@@ -4,7 +4,6 @@ Documentation for other versions
 Use the links below to access documentation for specific versions
 (when in doubt, use the **latest release**):
 
-* `Latest release <http://www.fatiando.org/tremelique/latest>`__
-* `Development <http://www.fatiando.org/tremelique/dev>`__
-  (reflects the *main* branch on GitHub)
-
+* `Latest release <https://www.fatiando.org/tremelique/latest>`__
+* `Development <https://www.fatiando.org/tremelique/dev>`__
+  (reflects the current development branch on GitHub)
