+++ environment.yml
@@ -1,47 +1,35 @@
 name: tremelique
 channels:
-    - conda-forge
-    - defaults
+  - conda-forge
 dependencies:
-    - python==3.12
-    - pip
-    # Build
-    - twine
-    - build
-    # Run
-    - numpy
-    - xarray
-    - numba
-    - matplotlib
-    - ipython
-    - ipywidgets
-    - imageio
-    - h5py
-    - rich
-    - ffmpeg
-    # Test
-    - jupyterlab
-    - pytest
-    - pytest-cov
-    - pytest-mpl
-    - coverage
-    # Documentation
-    - sphinx==7.2.*
-    - sphinx-book-theme==1.1.*
-    - sphinx-copybutton==0.5.*
-    - sphinx-design==0.5.*
-    - jupyter-sphinx==0.5.*
-    # Style
-    - black
-    - isort
-    - flake8
-    - flake8-bugbear
-    - flake8-builtins
-    - flake8-functions
-    - flake8-mutable
-    - flake8-rst-docstrings
-    - flake8-simplify
-    - flake8-unused-arguments
-    - pep8-naming
-    - pip:
-        - burocrata
+  - python==3.13
+  - pip
+  # Build
+  - twine
+  - python-build
+  # Run
+  - numpy
+  - numba
+  - h5py
+  - matplotlib
+  - ipython
+  - ipywidgets
+  - rich
+  - ffmpeg
+  # Test
+  - pytest
+  - pytest-cov
+  - pytest-mpl
+  - coverage
+  # Documentation
+  - sphinx==7.2.*
+  - numpydoc==1.7.*
+  - sphinx-book-theme==1.1.*
+  - sphinx-copybutton==0.5.*
+  - sphinx-design==0.5.*
+  - jupyter-sphinx==0.5.*
+  # Style
+  - ruff
+  - zizmor
+  - pip:
+    - burocrata
