+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "tremelique"
-description = "2D elastic wave simulator in Python"
+description = "Educational 2D elastic wave simulator in Python"
 dynamic = ["version"]
 authors = [
   {name="The Tremelique Developers", email="fatiandoaterra@protonmail.com"},
@@ -9,28 +9,34 @@ maintainers = [
   {name = "Leonardo Uieda", email = "leo@uieda.com"}
 ]
 readme = "README.md"
-license = {text = "BSD 3-Clause"}
-keywords = ["geophysics", "seismology", "waves"]
+license = {text = "BSD-3-Clause"}
+keywords = ["geoscience", "geophysics"]
 classifiers = [
-    "Development Status :: 5 - Production/Stable",
-    "License :: OSI Approved :: BSD License",
-    "Operating System :: OS Independent",
+    "Development Status :: 2 - Pre-Alpha",
     "Intended Audience :: Science/Research",
     "Intended Audience :: Developers",
     "Intended Audience :: Education",
-    "Topic :: Scientific/Engineering",
-    "Topic :: Software Development :: Libraries",
-    "Programming Language :: Python",
+    "License :: OSI Approved :: BSD License",
+    "Natural Language :: English",
+    "Operating System :: OS Independent",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.8",
     "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3.13",
+    "Topic :: Scientific/Engineering",
+    "Topic :: Software Development :: Libraries",
 ]
-requires-python = ">=3.8"
+requires-python = ">=3.9"
 dependencies = [
-    "pooch>=1.5.0,<2.0.0",
+    "numpy>=1.23",
+    "numba >= 0.53",
+    "h5py",
+    "matplotlib",
+    "ipython",
+    "ipywidgets",
+    "rich",
 ]
 
 [project.urls]
@@ -40,24 +46,85 @@ dependencies = [
 "Source Code" = "https://github.com/fatiando/tremelique"
 
 [build-system]
-requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
+requires = ["setuptools>=61", "wheel", "setuptools_scm[toml]>=8.0.3"]
 build-backend = "setuptools.build_meta"
 
 [tool.setuptools_scm]
 version_scheme =  "post-release"
 local_scheme =  "no-local-version"
-write_to =  "tremelique/_version_generated.py"
-
-# Make sure isort and Black are compatible
-[tool.isort]
-profile = "black"
-multi_line_output = 3
+write_to =  "src/tremelique/_version.py"
 
 [tool.burocrata]
 notice = '''
-# Copyright (c) 2024 The Tremelique Developers.
+# Copyright (c) 2025 The Tremelique Developers.
 # Distributed under the terms of the BSD 3-Clause License.
 # SPDX-License-Identifier: BSD-3-Clause
 #
 # This code is part of the Fatiando a Terra project (https://www.fatiando.org)
 #'''
+
+[tool.ruff]
+line-length = 88
+exclude = [
+    "doc/_build",
+    "src/tremelique/_version.py",
+]
+
+[tool.ruff.lint]
+extend-select = [
+  "ARG",      # flake8-unused-arguments
+  "B",        # flake8-bugbear
+  "C4",       # flake8-comprehensions
+  "D",        # pydocstyle
+  "EM",       # flake8-errmsg
+  "EXE",      # flake8-executable
+  "FURB",     # refurb
+  "G",        # flake8-logging-format
+  "I",        # isort
+  "ICN",      # flake8-import-conventions
+  "NPY",      # NumPy specific rules
+  "PD",       # pandas-vet
+  "PGH",      # pygrep-hooks
+  "PIE",      # flake8-pie
+  "PL",       # pylint
+  "PT",       # flake8-pytest-style
+  "PTH",      # flake8-use-pathlib
+  "PYI",      # flake8-pyi
+  "RET",      # flake8-return
+  "RUF",      # Ruff-specific
+  "SIM",      # flake8-simplify
+  "T20",      # flake8-print
+  "UP",       # pyupgrade
+  "YTT",      # flake8-2020
+]
+ignore = [
+  "ISC001",   # Conflicts with formatter
+  "PLR09",    # Too many <...>
+  "PLR2004",  # Magic value used in comparison
+  "RET504",   # Allow variable assignment only for return
+  "PT001",    # Conventions for parenthesis on pytest.fixture
+  "D200",     # Allow single line docstrings in their own line
+]
+
+[tool.ruff.lint.per-file-ignores]
+"__init__.py" = ["F401"]  # disable unused-imports errors on __init__.py
+"test/**" = [
+    "T20", # allow print statements in tests
+    "D",   # ignore pydocstyle warnings in tests
+]
+"doc/conf.py" = ["D"]  # ignore pydocstyle warnings in Sphinx config
+
+[tool.ruff.lint.pydocstyle]
+convention = "numpy"
+
+[tool.coverage.run]
+branch = true
+omit = [
+    "**/_version.py",
+    "**/__init__.py",
+]
+# Needed to combine coverage data from multiple OSs on CI
+relative_files = true
+
+[tool.coverage.report]
+show_missing = true
