+++ src/tremelique/_utils.py
@@ -1,9 +1,13 @@
-# Copyright (c) 2024 The Tremelique Developers.
+# Copyright (c) 2025 The Tremelique Developers.
 # Distributed under the terms of the BSD 3-Clause License.
 # SPDX-License-Identifier: BSD-3-Clause
 #
 # This code is part of the Fatiando a Terra project (https://www.fatiando.org)
 #
+"""
+Misc utilities that don't fit anywhere else.
+"""
+
 import numba
 import numpy as np
